[Master] Fix a bug when calculating the dispersion score
Summary: Previously I use the wrong type to keep the summary of the dispersion score.

Test Plan: Tested on migrate002

Reviewers: kannan, mbautin

Reviewed By: mbautin

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D451730

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1328560 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/master/AssignmentVerificationReport.java
@@ -212,8 +212,8 @@ public class AssignmentVerificationReport {
       }
     }
 
-    int dispersionScoreSummary = 0;
-    int dispersionNumSummary = 0;
+    float dispersionScoreSummary = 0;
+    float dispersionNumSummary = 0;
     // Calculate the secondary score for each primary region server
     for (Map.Entry<HServerAddress, Integer> entry :
       primaryRSToRegionCounterMap.entrySet()) {
